test(grid): move general test to component test file #3467
